# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org.quicko.qpon",
    "description": "Qpon is a coupon framework built by Quicko. Built in a scalable, multi-tenant thought process, the Qpon framework enables businesses to do campaigns and run offers using the Qpon API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{quicko_host}}"
    }
  ],
...
  "paths": {
...
    "/organizations/{organization_id}/coupons": {
      "post": {
        "tags": [
          "Coupon"
        ],
        "summary": "Create coupon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "coupon"
                  },
                  "name": {
                    "type": "string"
                  },
                  "discount_type": {
                    "type": "string",
                    "enum": [
                      "percentage",
                      "fixed"
                    ]
                  },
                  "discount_value": {
                    "type": "number"
                  },
                  "discount_upto": {
                    "type": "number"
                  },
                  "item_constraint": {
                    "type": "string",
                    "enum": [
                      "all",
                      "specific"
                    ]
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "{{organization_id}}"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Coupon"
        ],
        "summary": "Fetch coupons",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupons by specified status",
            "example": "inactive"
          },
          {
            "name": "discount_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupons by specified discount type",
            "example": "amount"
          },
          {
            "name": "external_item_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupons by specified item",
            "example": "item1"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to find coupon based on name"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/coupons/{coupon_id}": {
      "get": {
        "tags": [
          "Coupon"
        ],
        "summary": "Fetch coupon",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Coupon"
        ],
        "summary": "Update coupon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "@entity": "org.quicko.qpon.coupon",
                  "name": "Renamed coupon",
                  "item_constraint": "all"
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Coupon"
        ],
        "summary": "Delete coupon",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{coupon_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/coupons/{coupon_id}/campaigns/{campaign_id}/coupon-codes": {
      "post": {
        "tags": [
          "Coupon Code"
        ],
        "summary": "Create coupon code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "coupon-code"
                  },
                  "code": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "max_redemptions": {
                    "type": "number"
                  },
                  "minimum_amount": {
                    "type": "number"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  },
                  "duration_type": {
                    "type": "string",
                    "enum": [
                      "limited",
                      "recurring",
                      "forever"
                    ]
                  },
                  "expires_at": {
                    "type": "string"
                  },
                  "expires_after_billing_cycle": {
                    "type": "number"
                  },
                  "redemption_count": {
                    "type": "number"
                  },
                  "coupon-code-status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive",
                      "expired",
                      "redeemed"
                    ]
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Coupon Code"
        ],
        "summary": "Fetch coupon codes",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupon codes by specified status",
            "example": "inactive"
          },
          {
            "name": "visibility",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupon codes as per specified visibility",
            "example": "private"
          },
          {
            "name": "external_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Option filter to get customer specific coupon codes",
            "example": "customer@gmail.com"
          },
          {
            "name": "duration_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupon codes by specified duration type",
            "example": "recurring"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "56da8763-0a95-47ec-92f5-d8a2446d8a9c"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon set id",
            "example": "6e67d7cb-4fe0-4209-afe6-02ce07b5875a"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/coupons/{coupon_id}/campaigns/{campaign_id}/coupon-codes/{coupon_code_id}": {
      "get": {
        "tags": [
          "Coupon Code"
        ],
        "summary": "Fetch coupon code",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{campaign_id}}"
          },
          {
            "name": "coupon_code_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon code id",
            "example": "{{coupon_code_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Coupon Code"
        ],
        "summary": "Update coupon code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "@entity": "org.quicko.qpon.coupon_code",
                  "description": "₹300 discount, 1000 customers only",
                  "visibility": "private"
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon set id",
            "example": "{{campaign_id}}"
          },
          {
            "name": "coupon_code_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon code id",
            "example": "{{coupon_code_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/coupon-codes": {
      "get": {
        "tags": [
          "Coupon Code"
        ],
        "summary": "Fetch coupon codes by code",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "493c41aa-0311-4728-a2e0-3f051ca7bc8f"
          },
          {
            "name": "x-accept-type",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "To get response in paginated list format",
            "example": "application/json;format=sheet-json"
          },
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "{{x-auth-password}}"
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "INDMONEY20"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get coupon codes by specified status",
            "example": "inactive"
          },
          {
            "name": "customer_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Option filter to get customer specific coupon codes",
            "example": "customer@gmail.com"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/coupons/{coupon_id}/campaigns/{campaign_id}/coupon-codes/{coupon_code_id}/deactivate": {
      "post": {
        "tags": [
          "Coupon Code"
        ],
        "summary": "Deactivate coupon code",
        "requestBody": {
          "content": {}
        },
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API key for identification",
            "example": "{{api_key}}"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "JWT access token",
            "example": "{{access_token}}"
          },
          {
            "name": "x-api-version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "API version",
            "example": "{{version}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "coupon_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon id",
            "example": "{{coupon_id}}"
          },
          {
            "name": "campaign_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon set id",
            "example": "{{campaign_id}}"
          },
          {
            "name": "coupon_code_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Coupon code id",
            "example": "{{coupon_code_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
  }
}